+++ package.json
@@ -7,7 +7,7 @@
   "scripts": {
     "test": "echo \"Error: no test yet!\" && exit 0",
     "docs": "documentation readme dist/** --section=\"JavaScript API\"",
-    "precommit": "yarn run tsc && yarn run docs && git add README.md"
+    "precommit": "rm -rf dist && yarn run tsc && yarn run docs && git add README.md"
   },
   "repository": {
     "type": "git",
