+++ src/models/HitPayload.ts
